requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloud-sql-python-connector==1.2.4
cryptography==41.0.1
dominate==2.8.0
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozenlist==1.3.3
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-error-reporting==1.9.2
google-cloud-logging==3.5.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
migrate==0.3.8
multidict==6.0.4
pg8000==1.29.6
proto-plus==1.22.3
protobuf==4.23.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.17
typing_extensions==4.6.3
urllib3==1.26.16
visitor==0.1.3
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.9.2
